requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.2
cryptography==35.0.0
decorator==5.1.0
Elixir==0.7.1
et-xmlfile==1.1.0
ghostscript==0.7
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.20.1
idna==3.2
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.9
pandas==1.3.3
pbr==5.6.0
pdfminer.six==20201018
pdftopng==0.2.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabulate==0.8.9
Tempita==0.5.2
uritemplate==3.0.1
urllib3==1.26.7
webencodings==0.5.1
xlrd==0.7.1
xlwt==0.7.2
